+++ LatticeStatics/Utility/BFBSymPacCInterface.cpp
@@ -204,7 +204,7 @@ void InitializeOutputFile(char const* const datafile,char const* const startfile
 
 int RelativeEigVectsOK(Matrix const& EigVects)
 {
-   int const cutoff = 0.9; // 25.8 degrees
+   double const cutoff = 0.9; // 25.8 degrees
 
    int retval = 1;
    int size = EigVects.Rows();
